Change the dart-pad client to use the generated dartservices library.
I copied in the generated client stub library files to the dartservices_client folder. These files are generated and contains stuff that could be deleted. The nice thing is we don't have to maintain the client side stubs anymore.
Initially I had it as a local package which I referred to using a path dependency, but it seemed simpler and easier to maintain to just copy it in for now.
I do plan to change the generator to be able to just generate a single v1.dart file and have the common stuff as a package that can be depended on.
    Change the dart-pad client to use the generated dartservices library.

    Since last:
    . Added a mock version of the dataServices/v1 api.
    . Fixed the naming.

LMK what you think.

R=devoncarew@google.com, lukechurch@google.com
BUG=

Review URL: https://codereview.chromium.org/936053002