completion filter
This would solve both:
https://github.com/dart-lang/dart-pad/issues/152
and:
https://github.com/dart-lang/dart-pad/issues/202

I think this would also make the code of DartCompleter a little bit more logical. Not sure if you guys like it if I go this direciton, so before I move to fixing https://github.com/dart-lang/dart-pad/issues/203, I would like to hear if you guys agree with doing it this way.

Basicly, I added the class Completions, and I now use the suggestions of the analyzer about which text to replace.