Updated to appengine 0.6.0
 * This drops support for memcache.
 * Switched to use an in-memory cache instead.

It doesn't seem like this application was ever using a shared cache.
Instead each instance had it's own memcached instance running locally,
this makes no sense unless you're planning on migrating to a shared
memcached instance. As this is probably not needed, this commit switches
the application using a minimalistic in-memory cache.