add dark theme (#1075)
This adds new CSS files for light and dark themes, and
a separate frame for the dark theme in order to avoid flashing
the wrong background color. Uses a custom CodeMirror theme based on
zenburn that changes the background and gutter colors. adds a
moving_lines_dark.gif that is the original gif with inverted colors