Change how imports are parsed and evaluated.
This mostly simplifies how import directives are parsed, and how the URIs of
the imports are parsed and evaluated.

This allows multiple compiler errors to be returned if there are multiple
unallowed import directives. It paves the way a bit for multiple analysis
errors to be returned as well.

Work towards https://github.com/dart-lang/dart-pad/issues/1826.