Move Keyboard shortcuts into superclass, EditorUI.
With this new superclass, we can move the `Dialog dialog` field, a the common
key bindings, and showKeyboardDialog all up.

Each editor has some private implementation work with the remaining keybindings
(like `_format()` or `userCodeEditor`. While these may be deduplicated in the
future, it seemed a bit much for this change.

The superclass method `initKeyBindings` is `@mustCallSuper` to ensure that the
subclasses init the common keybindings as well as their custom ones.